+++ lib/rack/unpoly/middleware.rb
@@ -46,7 +46,6 @@ module Rack
       # Used internally by the middleware. Not required for normal use.
       # @api private
       def setup_protocol(request, headers)
-        headers["X-Up-Location"] = request.url
         headers["X-Up-Method"] = request.request_method
 
         if !request.get? && !request.env["rack.unpoly"].unpoly?